complete frontend operations

+++ app/todo/page.tsx
@@ -3,18 +3,15 @@
 import TodoInput from "@/components/TodoInput";
 import { useState } from "react";
 import { useItemsStore } from "@/lib/store/todoStore";
+import { uuid } from "drizzle-orm/gel-core";
 
-interface TodoObj {
-  heading: string;
-  description: string;
-}
+
 
 export default function Todo() {
   const [showButton, setShowButton] = useState<boolean>(true);
   const [showTodoInput, setTodoInput] = useState<boolean>(false);
-  const [todos, setTodos] = useState<TodoObj[]>([]);
-  const [heading, setHeading] = useState<string>("");
-  const [description, setDescription] = useState<string>("");
+ 
+  const [udpdateId, setUdpateId] = useState<number>(-1);
 
   //EVENT HANDLER
   function addTodo(): void {
@@ -22,46 +19,39 @@ export default function Todo() {
     setTodoInput(true);
   }
 
-  //delete todo function
-  function deleteTodo(index : number) : void {
-    const newTodos = todos.filter(todo => todos.indexOf(todo) !== index);
-    setTodos(newTodos);
-  }
-
-  //UPDATE TODO FUNCTION
-  function updateTodo(index:number) : void {
-    setTodoInput(true);
-  }
-   const { items ,deleteItem} = useItemsStore();
+ 
+  const { items, deleteItem, addItem, updateItem } = useItemsStore();
   // COMPONENT JSX
   return (
     <div>
       <h1 className="bg-black-200">this is todo page..!!</h1>
-      {/* {todos.length > 0 && (
-        <div>
-          {todos.map(({ heading, description }, index) => (
-            <div className="border border-blue-400 p-2 m-4" key={index}>
-            
-              <h3>{heading}</h3>
-              <p>{description}</p>
-              <p onClick={()=> updateTodo(index)}>UPDATE.!</p>
-              <span onClick={() => deleteTodo(index)}>&#10060;</span>
-            </div>
-          ))}
-        </div>
-      )} */}
+
       {/* NEWER VERSION */}
       {items.length > 0 && (
         <div>
-          {items.map(({ heading, description , id}) => (
-            <div className="border border-blue-400 p-2 m-4" key={id}>
-            
-              <h3>{heading}</h3>
-              <p>{description}</p>
-              <p onClick={()=> updateTodo(id)}>UPDATE.!</p>
-              <span onClick={() => deleteItem(id)}>&#10060;</span>
-            </div>
-          ))}
+          {items.map(({ heading, description, id }) => {
+            if (id === udpdateId) {
+              return (
+                <TodoInput
+                  key={id}
+                  heading={heading}
+                  description={description}
+                  onClickFunction={(head, descrip) => (
+                    updateItem(id, { heading: head, description: descrip }),
+                    setUdpateId(-1)
+                  )}
+                />
+              );
+            }
+            return (
+              <div className="border border-blue-400 p-2 m-4" key={id}>
+                <h3>{heading}</h3>
+                <p>{description}</p>
+                <p onClick={() => setUdpateId(id)}>UPDATE.!</p>
+                <span onClick={() => deleteItem(id)}>&#10060;</span>
+              </div>
+            );
+          })}
         </div>
       )}
 
@@ -73,13 +63,23 @@ export default function Todo() {
         </button>
       )}
 
-      {/* //TODO INPUT COMPONENT */}
+      {/* //TODO INPUT COMPONENT FOR NEW ENTRY*/}
       {showTodoInput && (
         <TodoInput
-          addButtonFunction={setShowButton}
-          todoInputBoxDispaly={setTodoInput}
-          addTodoFunction={setTodos}
+          description=""
+          heading=""
+          // onClickFunction={(head, descrip) => (
+          //   addItem({ id: Date.now(), heading: head, description: descrip }),
+          //   setShowButton(true),
+          //   setTodoInput(false)
+          // )}
+          onClickFunction={(head, descrip) => {
+            addItem({ id: Date.now(), heading: head, description: descrip });
+            setShowButton(true);
+            setTodoInput(false);
+          }}
         />
+        // />
       )}
     </div>
   );
